Added support for seting the location of compile_comands.json to a preset build directory.
LSP specific functions are added for clangd and ccls which should be called by the user by a
on_new_config fuction in their lsp setup function.

The changing cmake_soft_link_compile_commands is added as a config option which is
automatically disabled if one of the LSP functions is called.