Add hooks for clangd and ccls to get compile command directory from preset (#43)
* Removed hardcoded path for cmake

* Title: Disable soft linking of compile commands json file

Description:
The `cmake_soft_link_compile_commands` option was changed from `true` to `false
` to disable soft linking of compile commands json file. This allows for improved
 build directory management.

* Initial support for setting compile commands dir to preset build dir.

This adds a new flag const.cmake_compile_commands_from_preset which is
disabled by default. It hooks into the lsp settings and updates the
compile commands directory on successful build.

* Added support for seting the location of compile_comands.json to a preset build directory.

LSP specific functions are added for clangd and ccls which should be called by the user by a
on_new_config fuction in their lsp setup function.

The changing cmake_soft_link_compile_commands is added as a config option which is
automatically disabled if one of the LSP functions is called.

* Restored focus to previous window when opening cmake console

If a LSP is restarted when focused on the cmake output window
it will reattach to the output window buffer not the source buffer

* Restored original soft-link setting

* Removed debug statement

* Fixed usage of user settings

* Added usage to README

* Added missing configure_compile_commands() call