update configurePreset when changing buildPreset
A buildPreset can have a configurePreset associated with it. When
changing the buildPreset, the configurePreset should be updated
accordingly.

After updating the configurePreset, we have to re-generate using the
bang option as the new configurePreset might differ in the set
cacheVariables and we have to get rid of the cached variables of the
previous preset.