refac: restructure presets.helper function
The build_dir was overwritten if `binaryDir` was set. Moved it to the top with an early return to prevent the other computations

`set_dir_by_parent` now has a return instead of updating a surrounding variable. This also allows to iterate through the inherits
(parents) from the front and return once a dir is found.