Fix: build dir for preset with inherited presets without configurePresets (#189)
* fix: build dir for preset with inherited presets

A (build) preset might inherit from a preset without a configurePresets field set.
This causes p_preset to be nil when entering `helper`. The check for `not p_preset` does
not catch the unconditional access to `p_preset.name` resulting in an access to a nil variable

returning an empty string for a nil preset with the addtional nil check for the `ppreset` should
prevent the nil access

* refac: restructure presets.helper function

The build_dir was overwritten if `binaryDir` was set. Moved it to the top with an early return to prevent the other computations

`set_dir_by_parent` now has a return instead of updating a surrounding variable. This also allows to iterate through the inherits
(parents) from the front and return once a dir is found.

---------

Co-authored-by: Denzel <>