fix(terminal): failure to remove .lock file
When the runner/executor is set to `terminal` the removal of the .lock
file fails if the user has setup a shell alias for the `rm(1)` command that
requires confirmation (e.g.: alias rm='rm -iv').

In such case, the runner/executor does not finish and waits indefinitely
for user confirmation: remove ~/.local/share/nvim/cmake-tools-tmp/.lock?

This PR bypasses any alias of `rm(1)` command and also uses the `-f` flag.