feat: Add option to not pass --presets to cmake (#210)
## Changes Description

On CMake projects that use Conan as their package manager, there are two
presets that are created with the same name. This always causes any
`cmake` command issued by CMakeTools to fail. On those scenarios, the
project usually specify the CMake options (including the preset
location) as an argument to the `cmake` command.

This commit adds the option to not pass the `--preset` argument to the
`cmake` command.