fix: only generate after BufWritePost
This fixes an issue when the user has to select a configure preset while
generating the config.
If the generation is done at BufWritePre, the buffer is written while
the popup is still shown, which moves the cursor back to the now written
buffer. Selecting the popup to choose a preset is not possible anymore.

Writing the Buffer first and then prompting the user to select a
configure preset bypasses that issue