feat: improve runner and executer notification handling
With the new approach we get rid of the global states managed in the
notifications table. Everything is capsulated within the created
notification instance.
That also allows to fix an issue where a notification is dismissed while
a build process is still ongoing and the spinner should be updated.
Updating the spinner fails, because the notification timed out and got
dismissed, resulting in a repeating "No Matching notification to replace
found" error.
The new instance based approach tracks the shown notification and
replaces any existing notification when the message is updated. If the
notification times out and is dismissed, the instance will create a new
notification automatically.