Toggleterm: improve focusing and scrolling capabilities (#250)
* toggleterm: add scroll_on_error

add the option to only scroll when the operation failed. Together with
'auto_scroll = false' this is a nice alternative.

update the documentation for auto_scroll as it was slightly incorrect.
On long lasting operations you can notice that it scrolls an new input
received.

* toggleterm: add auto_focus

default it to true (this was the behavior all the time) but now it's
possible to disable it which would let the cursor stay in the editor.

* toggleterm: add focus_on_error

An addition similar to scroll_on_error. This lets the user select
auto_scroll = false and then the terminal will only get the focus when
an error occurred.

* dap: add initCommands