Add support for CMAKE_BUILD_TYPE presets using "type"/"value" form. (#297)
* Add support for CMAKE_BUILD_TYPE presets using "type"/"value" form.

With the current implementation, if CMakePresets.json sets
CMAKE_BUILD_TYPE as an JSON object with a "type" and "value" causes the
build step to fail -- example below.

```json
"CMAKE_BUILD_TYPE": {
  "type": "STRING",
  "value": "Debug"
}
```

This fix looks for a CMAKE_BUILD_TYPE with a "table" type and extracts
the "value" element, otherwise it falls back to the original behavior.

* Refactor to expand all cacheVariable in type/value form

* Had a rethink about this -- this will ensure expansion of vars in 'value'

* Use . syntax