check for valid window before resizing (#306)
when calling `notify` on a notification instance quickly after the
window has been dismissed, the on_close callback was not yet called as
it is handled asynchronously by vim-notify. Yet, the window is gone and
its winID is not valid anymore. Since the on_open callback is also
handled async, the winID has not been replaced by then. Skipping the
resize should be fine as the new notification should be displayed with
the correct size to begin with